# specs/002-role-interview-management/contracts/role-interview.openapi.yaml
openapi: 3.1.0
info:
  title: MockInterview Role-Based Interview Management API
  version: 0.2.0
  description: >-
    Executable feature contract for authenticated profile data, Candidate interview allocations,
    Manager-owned JDs and scheduling, shared Manager/Admin JD creation, and Admin user/JD pagination. All paths are relative to
    /api/v1. Authorization is server-enforced and every response carries X-Correlation-ID.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      summary: List all application users using deterministic server pagination
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Application-wide user page ordered by creation time then id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserPage"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{user_id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      operationId: getAdminUserDetails
      summary: Return available details for one selected application user
      responses:
        "200":
          description: Selected user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetails"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{user_id}/role:
    parameters:
      - $ref: "#/components/parameters/UserId"
    patch:
      operationId: changeAdminUserRole
      summary: Assign Candidate, Manager, or Admin and revoke the user's existing sessions
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role:
                  $ref: "#/components/schemas/Role"
      responses:
        "200":
          description: Role retained or changed; an actual change invalidates all old sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetails"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/job-descriptions:
    get:
      operationId: listAdminJobDescriptions
      summary: List all application JDs without confidential content
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Application-wide JD page ordered by creation time then id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJobDescriptionPage"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-mi_session
      description: Opaque HttpOnly Secure host-only application session cookie.
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
      description: Exact approved frontend origin; wildcard and missing origins are denied.
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
      description: Must match the readable CSRF cookie and server-side session digest.
...
  schemas:
    Role:
      type: string
      enum: [CANDIDATE, MANAGER, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    JobDescriptionSourceType:
      type: string
      enum: [MANUAL, UPLOAD]
    JobDescriptionFormat:
      type: [string, "null"]
      enum: [PDF, DOCX, TXT, null]
...
    PageMetadata:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    JobDescriptionSummary:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
    ManagerJobDescription:
      allOf:
        - $ref: "#/components/schemas/JobDescriptionSummary"
        - type: object
          required: [sourceType, sourceFormat, createdAt]
          properties:
            sourceType:
              $ref: "#/components/schemas/JobDescriptionSourceType"
            sourceFormat:
              $ref: "#/components/schemas/JobDescriptionFormat"
            createdAt:
              type: string
              format: date-time
...
    AdminUser:
      type: object
      required: [id, organizationId, displayName, email, role, status]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/UserStatus"
    AdminUserDetails:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/AdminUser"
        - type: object
          required: [profilePictureUrl, candidateProfileId]
          properties:
            profilePictureUrl:
              type: [string, "null"]
              format: uri
            candidateProfileId:
              type: [string, "null"]
              format: uuid
    AdminUserPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminUser"
    AdminJobDescription:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ManagerJobDescription"
        - type: object
          required: [organizationId, createdBy]
          properties:
            organizationId:
              type: string
              format: uuid
            createdBy:
              type: object
              additionalProperties: false
              required: [id, displayName]
              properties:
                id:
                  type: string
                  format: uuid
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 200
    AdminJobDescriptionPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminJobDescription"
    ErrorCode:
      type: string
      enum:
        - AUTHENTICATION_REQUIRED
        - SESSION_EXPIRED
        - SESSION_REVOKED
        - ACCOUNT_DISABLED
        - FORBIDDEN
        - CSRF_DENIED
        - VALIDATION_ERROR
        - RESOURCE_NOT_FOUND
        - IDENTITY_CONFLICT
        - IDEMPOTENCY_CONFLICT
        - UPLOAD_TOO_LARGE
        - UNSUPPORTED_DOCUMENT
    RecoveryAction:
      type: string
      enum: [SIGN_IN_AGAIN, CONTACT_ADMIN, RETRY, GO_TO_ROLE_HOME]
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, recovery, message, correlationId]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        recovery:
          $ref: "#/components/schemas/RecoveryAction"
        message:
          type: string
        correlationId:
          type: string
          minLength: 1
          maxLength: 100
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
  responses:
    InvalidRequest:
      description: Missing, malformed, past-time, blank, or otherwise invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthenticationRequired:
      description: Missing, expired, revoked, or invalid application session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Current role, Origin, or CSRF evidence does not permit the operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource unavailable; ownership and cross-tenant existence are not disclosed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Identity, role-transition, tenant-migration, or idempotency invariant conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...